[REF] core: refactor the `filtered_domain` method of `BaseModel`
The method filtered_domain() didn't conserve the order of `self`, which
may be confusing, and a subtle source of bugs.  It now preserves the
order from `self`.

The processing of `child_of`/`parent_of` making a search with the
default order of the model, which may be slow.

Add documentation.

Part-of: odoo/odoo#83687